pyproject.toml
[project]
name = "pvsonne"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexander Herrmann"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sonnenbatterie (>=0.6.1,<0.7.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
